Add .zshrc file and update readme to include ohmyzsh
This can be a bit confusing since ohmyzsh also creates a zshrc file and
we are also using the zshrc file to setup this repo. So once the install
of ohmyzsh completes we might need to re-add this repo to that file,
then we should be able to do a diff to see if the zshrc install did
anything un-expected.